Cleanup scheduler, fix process locks
Introduce the PROC_STATE_RUNNING state to signify process(es) that have
got a time slot to run. Manage that state in the scheduler.

Move the round-robin selection of the next process out into a separate
function for readability, make it look specifically fora process with
PROC_STATE_READY.

Acquire and release locks on proc and last_proc within scheduler.