UART reading, basic read() syscall and a (not yet functional) shell
Add trivial UART reading. Implement read() syscall, which only reads
from UART for now.

With all that, add a trivial shell, but it works funky for now because
we're missing a wait() syscall.

qemu-launcher.py will need a lot of improvement to support input passing
to qemu. For now, I disabled input piping, so quitting now requires qemu
console: C-a c, quit.

I was also using cat until now for reading from HiFive on /dev/ttyACM0,
but that obviously doesn't do input. Unfortunately, the regular screen
fails to translate newlines for me, so I've added picocom to the list of
packages, as well as scripts/term.sh to launch it with required params.