Implement userland printf()
It's called uprintf() for now, due to naming conflict. Will clean up
later.

It has an unfortunate usage quirk: any string parameters must be
declared not inline, and with _user_rodata attribute when global, in
order to place it in userland-accessible memory. Otherwise the string
gets info the regular .rodata and we get PMP access violation when
trying to read it.